Add files via upload
Back in 10th grade, during my dear friend's birthday, I whipped up a sketch for her. She was a huge T.S. fan, so I put in extra love! After exams, I dedicated a solid couple of hours to it. Started with a rough tracing of the real image, then dove into adding details one by one. And get this - it's all done with just a pencil! I even made a copy for myself. Stay tuned for the upload!