Add files via upload
On December 25th, I embarked on the task of creating an invitation card for the Colonel of 55 UP BN NCC, marking the occasion of Republic Day 2023. With the assistance of an olive-colored sheet borrowed from a former friend and ivory sheets from my own supplies, I set to work. Drawing inspiration from a captivating idea discovered on Pinterest, I meticulously sketched, cut, and assembled each component with great care. Crafting the Colonel's attire, meticulously detailing the blazer and medals, I used a marker to accentuate the edges, ensuring every detail stood out. With the aid of a fevistick, I brought the pieces together, culminating in the final product you see before you. The entire process was a delight, each step enhancing the project's beauty and intricacy. I trust you will derive as much pleasure from viewing it as I did from creating it.