TeddyCard
Recently, I found myself lacking motivation for creativity, feeling a significant absence of desire to draw or engage in any artistic endeavors. However, with the upcoming birthday of a dear friend approaching, I felt inspired to create something special. Drawing inspiration from a Pinterest account called 'crafftitlover', I decided to handcraft a greeting card for her in advance. Using red and ivory-colored sheets of paper, I meticulously sketched and assembled the components: the teddy bear's head and hands, as well as a heart. Employing double-sided tape, I brought these elements together to achieve a delightful 3-D effect. To accentuate the details, I used a black marker along the edges and added a touch of blush using a red brush pen. The caption, inspired by 'Lusan Huitema' on Pinterest, was penned in white gel ink. Additionally, I incorporated a pre-made 3-D sticker of a blue heart, left unattached to convey the impression that the teddy bear is holding it, allowing for flexibility in its placement.